+++ lib/open_exchange_rates.ex
@@ -9,18 +9,27 @@ defmodule OpenExchangeRates do
   def start(_type, _args) do
     import Supervisor.Spec, warn: false
 
+    configuration_status = check_configuration
+
     children = [
       worker(OpenExchangeRates.Cache, []),
     ]
 
-    children = case check_configuration do
+    children = case configuration_status do
       :ok -> children ++ [worker(OpenExchangeRates.Updater, [])]
-      :error -> children
+      _ -> children
     end
 
-    opts = [strategy: :one_for_one, name: OpenExchangeRates.Supervisor]
-    Supervisor.start_link(children, opts)
+    opts = [strategy: :one_for_all, name: OpenExchangeRates.Supervisor]
+    {:ok, pid} = Supervisor.start_link(children, opts)
 
+    case configuration_status do
+      :disable_updater -> load_data_from_file
+      :missing_key -> load_data_from_file
+      _ -> nil
+    end
+
+    {:ok, pid}
   end
 
   @doc"""
@@ -47,11 +56,11 @@ defmodule OpenExchangeRates do
   @spec convert(Float.t, (String.t | Atom.t), (String.t | Atom.t)) :: {:ok, Float.t} | {:error, String.t}
   def convert(value, from, to) when is_float(value) do
     with \
-      {:ok, rate_to_usd} <- OpenExchangeRates.Cache.rate_for_currency(from),
-      {:ok, rate} <- OpenExchangeRates.Cache.rate_for_currency(to) \
+      {:ok, rate_from} <- OpenExchangeRates.Cache.rate_for_currency(from),
+      {:ok, rate_to} <- OpenExchangeRates.Cache.rate_for_currency(to) \
     do
-      usd = value / rate_to_usd
-      converted = usd * rate
+      rate_usd = value / rate_from
+      converted = rate_usd * rate_to
       {:ok, converted}
     else
       error -> error
@@ -75,23 +84,29 @@ defmodule OpenExchangeRates do
     end
   end
 
-
   @doc """
-  Get the conversion rate for a given currency to USD"
+  Get the conversion rate for a between two currencies"
 
   ## Example
 
-      iex> OpenExchangeRates.rate_for_currency(:EUR)
-      {:ok, 0.902}
+      iex> OpenExchangeRates.conversion_rate(:EUR, :GBP)
+      {:ok, 0.8481186252771619}
 
   """
-  @spec rate_for_currency((String.t| Atom.t)) :: {:ok, Float.t} | {:error, String.t}
-  def rate_for_currency(currency), do: OpenExchangeRates.Cache.rate_for_currency(currency)
-
+  @spec conversion_rate((String.t| Atom.t), (String.t| Atom.t)) :: {:ok, Float.t} | {:error, String.t}
+  def conversion_rate(from, to) when is_binary(from) and is_binary(to), do: conversion_rate(String.to_atom(from), String.to_atom(to))
+  def conversion_rate(from, to), do: convert(1.0, from, to)
 
   defp check_configuration do
-    case Application.get_env(:open_exchange_rates, :app_id) do
-      nil -> Logger.warn ~s[
+    cond do
+      Application.get_env(:open_exchange_rates, :auto_update) == false -> :disable_updater
+      Application.get_env(:open_exchange_rates, :app_id) == nil -> config_error_message; :missing_key
+      true -> :ok
+    end
+  end
+
+  defp config_error_message do
+    Logger.warn ~s[
 OpenExchangeRates :
 
 No App ID provided.
@@ -99,14 +114,23 @@ No App ID provided.
 Please check if your config.exs contains the following :
   config :open_exchange_rates,
     app_id: "MY_OPENEXCHANGE_RATES_ORG_API_KEY",
-    cache_time_in_minutes: 1440
+    cache_time_in_minutes: 1440,
+    cache_file: File.cwd! <> "/priv/exchange_rate_cache.json",
+    auto_update: true
 
 If you need an api key please sign up at https://openexchangerates.org/signup
 
 This module will continue to function but will use (outdated) cached exchange rates data...
-      ]
-      :error
-      _ -> :ok
+    ]
+  end
+
+  defp load_data_from_file do
+    if File.exists?(OpenExchangeRates.Cache.file) do
+      OpenExchangeRates.Cache.file
+      |> File.read!
+      |> Poison.decode!
+      |> Map.fetch!("rates")
+      |> OpenExchangeRates.Cache.update!
     end
   end
 end